Tracks now stay in sync when changing the tempo during playback
Let's say you have two Tracks:

C1 - C1 - C1 - C1 -
A3 - -  - A3 - -  -

Previously, if you changed the tempo during playback, the two
Tracks could become out of sync with each other (i.e. the C1
and A3 would not be played together at the same time).

To fix this, the parsing of each Track sequence has been changed.

Previously, "A3 - - -" would be parsed as a single note with a
duration of 4 steps. Now, it will be parsed as a note with duration
of 4 steps, and 3 rest notes with a duration of 1 step each.

The benefit is that Track can simply iterate through its set of
notes one at a time, and always have a note to play at each step. No
special logic is needed during playback to account for the fact that
different notes might have different durations.

This seems sort of hacky, so not sure how well this will hold up
long term. However, it solves the immediate problem of being able
to change the tempo without Tracks becoming out of sync.