Extracting Pattern object for a set of Tracks that play at the same time
In particular, when added/replacing a Track during playback, allows the
new Track to be synced with the other Tracks. This happens because there
is now a single 'current time' shared between all of the Tracks, instead
of each Track tracking their own times (which are supposed to be in sync,
anyway).