A Track's sequence is no longer mutable
When a sequencer track's notes are changed in the UI, a new
Track is now created to replace the old Track, instead of mutating
the original Track. This allows consolidating the logic of parsing
a raw string into a Sequence, and is also in favor of generally
trying to favor immutability.