Making Instrument instances AudioContext agnostic
An AudioContext is now provided to Instrument.playNote(), instead
of being given to the Instrument upon construction. This allows
an Instrument to be constructed before the appropriate AudioContext
that it will be used with will exist.

This will be especially important for offline audio rendering,
where the length of the output buffer must be defined when the
OfflineAudioContext is constructed. If Instruments rely on the
OfflineAudioContext to exist when they are constructed, this
creates a circular dependency where both rely on each other in
order to be constructed (unless the OfflineAudioContext is given
a one-size-fits-all buffer length, causing extra silence to be
present at the end of the buffer).