Preventing clipping in downloaded Wave file
Tried using a Compressor node, but couldn't get the output to match
what it sounds like when playing back in the browser. For example,
the attack on some notes was being made longer.

Instead, this approach manually scales all samples based on the
master volume to prevent any samples from being over 1.0.