Fixing audio playback on initial page load in Safari
The comment in the code change essentially describes
what this is fixing. For some reason, when initially
loading the page in Safari 9.1 (and maybe 9, although
I haven't tested in that version), the AudioContext
used for playback is in suspended state, and thus
won't generate any sound. If you re-load the page,
it will no longer be in suspended state and audio
playback will work.

This fix works around that, so that audio playback
properly occurs in Safari.