Playback now occurs via a sequence of Patterns
Instead of via playing a single defined Pattern. This is the
first step of allowing multi-Pattern sequencing to support
playing a full song, not just a single Pattern.

This is accomplished via the new SongPlayer object. This
represents a sequence of Patterns to be played, along with
the state of where in playback the player is.

Currently, only a single Pattern can be played at a time. At the
moment this is just to simplify the initial implementation.