New sequence parsing implementation with simpler algorithm
It scans the input backwards instead of forwards. This reduces
the amount of state that needs to be remembered while traversing
the input.

Previously, when a note was encountered, it wasn't
known how long the note would be without scanning forward in the
input, so the note details had to be remembered so they could
be added at a later time when the duration was known.

Now, due to scanning backwards,  any time a note is encountered,
it can be added on the spot since the duration is already known.

This also removes the need to handle edge cases such as adding
an 'in progress' note if the end of the input is reached, etc.