Avoiding side effect when calculating note frequency
I.e., a note's name is not converted to its enharmonic
equivalent during construction. Although a normalized
note name is used to simplify calculating the frequency,
the incoming note name is no longer fully replaced with
this normalized name. Since I don't have a clear reason
for why a note's name should be replaced with a normalized
equivalent, I'm going with just preserving what is passed in.

Also fixing incorrectly specified tests in the process