Sequencer display stays in sync after tempo change during playback
I.e., before this commit the current measure was correctly
highlighted during playback, as long as the tempo wasn't
changed during playback. If the tempo _was_ changed, then
the sequencer display would go haywire, and the measure
highlighted in the UI would not necessarily be the measure
actually being played.

This commit fixes this by using the technique described in
section "Yet Another Timing System" at:
https://www.html5rocks.com/en/tutorials/audio/scheduling/