Preventing busy polling after sequencer playback is stopped
The timeoutId variable was inadvertantly being shadowed,
which meant when an attempt was made to stop the timeout
interval, the wrong timeoutId was used. This caused the
interval to be fired indefinately after playback stopped,
until the page was reloaded.

While fixing this, also adding a change to manaully
null out the current sequencer step when playback stops,
so that the 'current measure' highlight is correctly removed
from the measure that was being played when playback stopped.
Otherwise, it will stay highlighted indefinately. The
previous busy polling interval inadvertantly caused the
highlight to correctly be removed.