Sequencer no longer has concept of "currently selected track"
As a result, when adding a new track, the track that is
selected in the track editor does _not_ change to the
newly created track. After playing around with it, this
seems like the right call - just because you add a track
doesn't mean you necessarily want to edit it right away.