Attempting to bind properties to controller, instead of using scope
Not totally sure I'm doing this right... but functionally the
app still seems to work correctly.

The intended purpose is to try to get the app to be closer
to being component based.

Even if this isn't fully correct, it does seem to result in
less code, and fewer messages being emitted...