Preventing error when removing track
Caused by a race condition with the song being serialized in
reponse to multiple events, resulting in the removed track
still existing, but having an instrument ID for an
instrument that no longer exists.

Currently, there aren't specific actions that should be taken
when adding or removing an instrumetn, so removing these event
messages.