Changes to are reflected more quickly during playback
They are now reflected as soon as the next notes are
scheduled - previously they were only reflected
once the next pattern boundary started (i.e. every
16 steps).

The work to do this also simplifies the core synth,
because the playback logic now doesn't need to know
about patterns or tracks - it only knows that certain
notes need to be played at certain steps.

As a result, the concept of "pattern" and "track"
and the associated code has been removed from the
core synth.