Preventing error when removing the only sequencer pattern
The error was caused by using the number of patterns defined
in the first sequencer track to determine the length of the
song. However, when removing the only track, there will be
no first track, causing the error.

To fix this, adding a field to SequencerService to indicate
the total length of the song (in steps). Currently, this is
hard-coded to a constant value, but this allows a more direct
way to query this value, and makes it easier to allow
changing the song length in the future.