Preventing pops and clicks during playback
It appears to be caused by suddenly setting the gain
volume to 0 at the same time as note playback starts.

To fix this, the gain volume is set to 0 at time 0,
so that it is guaranteed to already be at 0 before
the note starts playing.

This works because a new gain node is constructed
each time a note is played, so it doesn't matter
_when_ the gain volume was set to 0, as long as it
happens before the note starts playing. By setting
the volume to 0 at time 0, then this will occur
for any note (except for notes that start at time
0, but those don't seem to have a problem with
pops/clicks).

Also, popping/clicking seemed to be browser
dependent - noticeable in Safari and Firebox,
less so in Chrome.