Removing ability to mute pattern rows
Now that the song can have multiple tracks, each track having
its own instrument and patterns, the ability to mute a single
row in a pattern seems superfluous. Not sure why you would
want to mute a pattern row in practice, so removing the
feature. Can always re-add it later.