Adding filter envelope to instruments
The filter can now either be modulated by the LFO
(the existing "wobble" option), or via an envelope.
Depending on what mode is chosen, the corresponding
LFO or envelope controls are shown in the UI.

Some combinations of envelope+filter settings seem to
cause pops/clicks, so those might need to be fixed in
a future commit.