Better default names for new patterns
Previously, was "Pattern <n>", where <n> was the internal
ID of the pattern. This would result in strange names
such as "Pattern 18" when adding a new pattern to the
demo song.

Now new pattern names are in the form "<instrument name> <n>",
where <n> is the number of patterns for the current instrument
plus one. If you delete patterns this can cause new patterns
to have the same name as a previously created pattern. Maybe
that can be fixed in a future commit, but for the time
being this is still an improvement over the previous behavior.