Reducing track volumes to reduce change of clipping
The max volume per track is now 1 / num_tracks. This way, if
all tracks are playing a single note at once at maximum volume,
there shouldn't be clipping in theory.

This isn't perfect - for example, if a track plays more than one
note (such as a chord), it could still cause clipping to occur.
It also means that the song gets quieter as more tracks are
added. There is probably a better way to handle this, but it
is at least better than it was before - in Safari there was
very obvious clipping for the demo song at 50% master volume.