Fixing export to *.wav file
Apparently it's valid to set a paramater value at a time
less than 0 for an online AudioContext, but not valid
for an offline AudioContext. This was causing the export
to *.wav functionality to generate an error for the demo
song.

To fix this, clamping parameter start times to be 0
or greater.