Adding alternate version of app written using React
This is essentially a re-write of the UI layer. The reason
for doing this is to allow moving on from AngularJS, now that
the writing is on the wall that it won't be supported long
term. React increasingly seems like a good choice, and the
new Angular versions look unpleasant to use. The goal is to
delete the old AngularJS implementation in a future commit
and replace it with this new version.

Currently, the entire React implementation is in a single
file. This was done to make it easier to convert the app
without having to deal with other concerns like WebPack and
module loading yet. In the future, this giant file should be
decomposed.

It's also using just-in-time in-page Babel translation for
ES6 code, this should also be moved to an ahead of time step
in the future.

Finally, there are undoubtedly ways the new React code could
be cleaned up, made more idiomatic, etc. However, this is a first
pass, and converting the entire app to React is a big step!