Pressing keyboard key plays note using current instrument
This hooks up basic "press a key and hear a note" functionality,
but still limited. Notes always play for a constant duration, no
matter how long the key is held, and play at a loud volume.

However, even in this basic form it is still fun to play around with.