New implementation of keyboard touch controls
Although more complicated, and seemingly less "React-like", this
implementation allows touch controls to work in a more sensible
manner. For example, you can now drag your finger across the keyboard
to play a glissando.

The reason for this implementation is that the normal touchmove
event doesn't fire on the element that the finger is currently over,
but rather the element on which the touch event started. This means
that unlike with mouse events, you can't add a touchmove event to
each key and trigger the key if it fires, because these events won't
actually fire as you drag your finger around.

To make the implementation simpler, mouse event are handled in the
same way as touch events, even though for a mouse-only keyboard
there would be a simpler way to implement this. But when handling
both mouse and touch controls, it seems simpler to have one standard
way of handling events.