Scheduled sample instrument notes stop at end of note duration
Previously, sample instrument sounds would play for the entirety
of the sound sample. If the sound had a shorter duration than the
note this would sound OK, but if the sound was longer it would
start overlapping with other notes.

Now, the sound will only play for the duration of the note.