General refactoring of logic to play a synth instrument note
The biggest change is changing the order that the relevant WebAudio
nodes get created. Creating the nodes "backwards", i.e., the nodes
that are closest to the final destination node first, followed by
nodes further away, results in more straightforward code. The reason
is that any time node A needs to connect to node B, node B already
exists. This results in code that is more "straight line" logic of
"create node, then connect it". Otherwise, the logic is "create node,
do some other stuff, then connect it".