Newly created tracks are automatically selected in track editor
My experience of using the app is that after creating a new track
I usually want to immediately edit it, so this behavior is convenient.
I've also sometimes accidentally edited the wrong track, since I
incorrectly assumed that I was editing the newly created track, and
in theory this will make that less likely.