Transport properly synced after new sampler track file uploaded
Previously, the transport would not play a sampler track's newly
uploaded sound file until an action was taken to manually sync
the transport (such as mute/un-mute a track). The reason was the
an attempt was made to sync the transport when uploading the file,
but this would occur before the file had completed loaded and thus
not take effect.

Now, the sync is delayed to not happen until the file has finished
uploading.