File upload form shown/hidden based on toggle
Previously, it was always present on the screen. Now, it is only
shown when the user clicks a link indicating they want to change
the sound file, and disappears once the file is uploaded.

This seems like a better UI, since the file upload form isn't
relevant unless you actively want to change the sound file.

This does not have the correct visual layout yet, so that will
need to be fixed in a future commit.