Explicitly set root font size
This shouldn't make any changes to the visual appearance, but
makes more explicit that the <rem> units used all over the place
are based on 16px, not the 20px font size set on the <body> tag.

This difference is confusing, and will hopefully be rectified in
the future. For now, this helps make it a little more explicit.