Better UI for changing sound file for a sampler track
Clicking the "change" link now automatically causes the file chooser
dialog to be shown. Canceling the dialog cancels the file upload, and
choosing a file causes it to be automatically loaded.

This is a simpler UI from the user's perspective, because you no longer
need to manually confirm the upload, there is a more obvious way to
cancel the upload, and the "no file selected" error can no longer occur.

This is also simpler from an implementation perspective, because the
dynamically shown/hidden upload form no longer needs to exist.