User prompted for sound file before creating sampler track
When the user clicks the "Add Sampler Track" button, they are
now immediately shown a file chooser dialog. If they select a
valid sound file, a new sampler track will be created using
that sound file. If they click cancel, the track won't created.
If they choose an invalid sound file, they'll see an error
message, and the track won't be created.

The benefit of this approach is that a default placeholder sound
never needs to be set on the track - it guarantees that a real
sound chosen by the user is always present. In practice, it
doesn't make sense for a sampler track without a sound sample to
exist, and this approach guarantees a sample will always be
present.