Removing unnecessary logic
The visibility of the `PatternEditor` component is controlled by
either including the component in the DOM or not, so this additional
method of showing/hiding via CSS is redundant. I think this is a
relic of the old Angular implementation of the app that can be moved
now that the app is implemented using React.