Instrument fields collapsed into tabbed interface on mobile
This seems to be a better UI for mobile, as it avoids a very long
amount of content for the instrument editor, and makes it easier
to tell where you are. It also makes it a little easier to change
instrument parameters while also using the on-screen keyboard.

On wider viewports, the instrument editor has the same apperance
as before - all fields are shown at once in a horizontally scrolling
pane.