Page content doesn't overflow on mobile screen size
Previously, the content of the transport tempo and volume controls
overflowed slightly on narrow viewports, causing the page to
be horizontally scrollable. This caused issues with touches on
the page.

This doesn't _fully_ fix the issue - in iPhone SE mode in Safari
developer tools, it still overflows, but doesn't at the iPhone 8
screen size. So, it is an improvement at least.

This causes a regression in the tablet viewport size, since the
two controls now expand to fit the window, instead of each being
50% width side by side. That should be fixed in a future commit.

Despite the regression, this seems like a step in the right
direction in the bigger picture.