Refactoring to make public/private distinction more clear
The returned object at the end of the constructor function shows
which fields are "public" at a glance, instead of having to look
back through the rest of the constructor function.

This is only being done to objects in the core synth code, and not
every object is being converted at the present time.