Preventing pops/clicks in audio saved to *.wav file
When generating the saved output, the sample data is inspected to
determine the maximum sample value, to determine how the final output
should be scaled. However, this was previously only taking positive
samples into account, and not negative samples. If the largest negative
sample value was larger than the largest positive value, this could
cause the output audio to be scaled by the wrong amount and become
clipped.