Left/right keyboard scroll buttons can be pressed at same time
Previously, this would cause the two buttons to be unresponsive
until they were no longer touched. Now, they work as you would
expect: if both are pressed, then they cancel each other out, and
if one of the buttons is released the keyboard will start scrolling
in the direction of the button still being pressed.