More detailed info about song position during playback
Now the current step is shown, in addition to the current measure.
The current step is shown by a sort of "playback head". This opens
up the possibility for starting playback at any arbitrary position
in the future, by either manually moving the playback head, or
starting from the position the playback head was at when the sequencer
was last stopped.

As part of the implementation, this fixes the width of each measure
in the sequencer, instead of making it expand to fill the window.
In addition to helping the current implementation, this seems like
it will help when allowing songs to have an arbitrary number of
measures.