Better appearance for extra background space in sequencer
I.e. the space between the last measure and the right edge of the
viewport, if the viewport is wider than the full set of measures.
Previously, this was just a white void, but now it has a light
gray background to tie it more closely with the rest of the sequencer
visually.