Pressing letter key sets note name regardless of cursor position
This makes it easier to enter notes, by making the note inputs less
fussy about where the cursor currently is.

This makes it impossible to enter flat or double-flat notes, since the
"B" key now always sets the note name. This is intended to be a
temporary reversion, to be fixed in a future commit.