Note can be made flat by pressing "@"
This is not necessarily super discoverable, but avoids the context
issues that come from using "B" to mean either "note name B", or
"flat". It seems a little logical because it is to the left of the
"#" key, and "@" doesn't seem to have another obvious use.

The amount of flatness can be cycled by pressing "@" multiple times.

This has a known issue where hitting "@" before a note name is given
will cause the note name to be set to "B", instead of adding a flat.
This should be fixed in a future commit.