Onscreen keyboard doesn't cause current pattern note to lose focus
This makes using the onscreen keyboard to enter pattern notes much
easier to use. For example, you can select a note with the mouse,
press a key on the onscreen keyboard, and then move to the next note
by pressing the left arrow key. This works because using the onscreen
keyboard doesn't cause the current pattern note to lose focus.

This also provides a simpler model for tracking the state of which
note has input (if any). It is now explicitly tied to the note input
having focus, which seems like a more logical UI.