Preventing text keyboard overlay for note inputs on iOS
This makes note entry on iOS at least sort of usable, by forcing
all note entry to happen via the onscreen musical keyboard. Previously,
the text keyboard would block most of the screen, making note entry
impractical (at least on a phone size device).

This is not a fully usable solution, because it prevents any way
of erasing notes or elongating notes with "-" on iOS. That will
need to be fixed in the future, but this commit allows testing out
this new approach.

Not sure if this works the same way on Android.