Refactoring to use props as data source, not <input> value
This was a leftover from the Angular -> React conversion. Using the
passed in prop value makes the code simpler, and I think easier to
understand. It's also more React-like. In particular, it removes the
need to translate back and forth between "formatted note" and
"unformatted note" formats. Instead, there is now a one-way flow from
unformatted -> formatted notes values.