Quiet audio not artificially amplified in downloaded *.wav file
Previously, all downloaded audio was scaled so that the largest
sample value was 1.0. This worked as a way to bring down loud audio
to prevent clipping, but also had the effect of making quiet audio
louder. In fact, it meant that the final audio would always be the
same loudness, regardless of where the volume sliders were set in
the app. (Except when the volume was 0; in that case the output
volume would correctly be 0).

Ideally, the online and offline outputs should go through the same
audio path and not have separate amplitude scaling, but in the
immediate term this commit improves the situation.