Experimental filter envelope; relative to cutoff, instead of 0
That is, the filter now starts at the cutoff frequency, goes up to
a certain amount, sustains at a certain amount, and finishes at the
the cutoff frequency.

Previously, the filter started at 0, when up to the cutoff frequency,
sustained at a certain amount, and finished at 0.

The new design seems a little more flexible to use, and also makes it
less likely to experience pops/clicks during the release portion of
the envelope. It also allows simplifying the core synth code a little
bit.