Adding ability to calculated envelope value at a given time
The intended use case is to enable using an envelope for interactively
trigged notes (i.e. from the on-screen keyboard), and having the
envelope work correctly in all browsers. The current problem is that
triggering the release portion of an interactive note requires knowing
what the current envelope value is, and as far as I know there doesn't
seem to be a cross-browser way to do that other than calculating it
manually. (cancelAndHoldAtTime() _would_ work, but it's not supported
widely at the time of this commit).