Fixing interactive envelope release in Firefox
Previously, when releasing a keyboard key in Firefox, the volume
of the note spiked to full volume, and decayed from there. The
reason was that Firefox doesn't seem to reflect gain changes from
functions such as `linearRampToValueAtTime()` in the gain node's
gain value, while Chrome and Safari do.

The solution is to manually calculate what the gain should be at
the time the key is released, and use that for setting the initial
gain value of the release.