Disabling *.wav download if no 'download' attribute support
For example, as of this commit, Mobile Safari doesn't support the
`<a download="">` attribute. This causes downloads to open in a
new tab as a blob, instead of actually downloading a file. In
practice, the behavior is wonky and doesn't seem to always work
correctly.

Rather than have inconsistent behavior of questionable value
(since you can't actually download the file), disabling the feature
altogether unless the browser has full support for downloads.