Unambiguous representation of a flatted note ("@" instead of "b")
This makes parsing a little simpler, and fixes the bug in which typing
"@" into an empty note field would display as "B" instead of the flat
character. Although notes look a little weird with the "@" character,
it doesn't seem that confusing and the benefit on an un-ambiguous note
format seems worth it.

This only reflects the internal data format, and not what is displayed
in the UI.