Refactoring pattern editor to rely on single note input
Each cell in the pattern grid is only for display, and is not an
<input> that accepts key presses. Instead, there is a single
invisible <input> that handles all key presses, and is focused as
needed when a pattern cell is clicked.

The benefit of this is that input handling and presentation can be
split into separate concerns. This also avoids the need to style
<input> elements, instead of plain <span>s.