Experimentally using San Francisco as app font, not Helvetica Neue
A benefit is that it allows text to be rendered at the correct weight
in Chrome, due to the Chrome's apparent bug with rendering Helvetica
Neue at 300 weight. (It renders the text too thin, making it hard to
read). Not necessarily the best reason for changing the app font, but
maybe pragmatic.