Pattern names automatically based off track name, can't be edited
In practice, I found that editing the pattern names and keeping them
in sync with the parent track name was annoying, and I didn't find
myself wanting to give patterns custom names. It also led to bad
experiences, such as adding a new track with default name "Synth Track",
changing the track name, but then seeing the default pattern still
having a name of "Synth Track 1".

A drawback of this approach is that it can be confusing if a pattern
name changes automatically because of e.g. removing another pattern.
It could make it more difficult to remember what "Pattern 2" represents
over time. Will try the app out like this and see how big of a problem
it is in practice.

It seems like the _real_ solution to this is to have a more flexible
sequencer where you can have arbitrary pattern placement, and patterns
can be identified visually instead of by name. However, that's a much
bigger change, so this is an intermediate improvement.