Pitch LFO wobbles same relative amount, regardless of note
This is accomplished by tying the LFO to the oscillator detune,
instead of frequency. For example, if setting the detune to 700
cents, the LFO will move between a fifth above and a fifth below,
regardless of the root pitch. When setting the raw frequency, the
amount of change it sounds like is relative to the root frequency
(applied to a low frequency the LFO will sound like a bigger change,
and applied to a higher frequency it will sound like a smaller
change). A constant amount of change regardless of the root pitch
is probably what you normally want.