Filter cutoff LFO amount consistent regardless of cutoff frequency
This is a companion to 3ab6f30f91f420701006eb074872d6370a4fec9d.
Similar to how the pitch LFO amount was made to be the same regardless
of the underlaying note frequency in that commit, the filter LFO
amount is not the same regardless of the cutoff frequency. This is
due to modulating the filter cutoff detune value, instead of the
filter cutoff frequency.