Transport play button appearance more consistent across browsers
Now that it is rendered via CSS instead of an actual glyph, the
appearance should not depend on the fonts that the user has install
on their computer. In particular, the triangle in the button should
always have a fixed size, and the button should always appear as a
circle, not an ellipse.