Making accidental "swipe left to go back" gesture when using slider
When a slider's value was 0, grabbing onto the the slider thumb and
dragging it could accidentally trigger the "swipe left to go back"
gesture on iOS Safari. Although the smaller margins look better and
allow more content to be visible on the screen, it's even more
important for the slider controls to be usable.