Better experience if loading error due to no WebAudio API
Previous, a blank screen was shown. Now, an error message is displayed
to give a nicer indication of why the app can't load.

This allows simplifying the transport component a little bit, by removing
the ability to show an error message there while allowing the rest of
the page to function. Now that the app is more complex, there's more
functionality that would also be broken in addition to the transport
(sample loading, keyboard). It might be nice in the future to allow the
app to function but without audio, but that's more complicated and seems
like an edge case since the WebAudio seems to be pretty widely supported
as of this commit.