Muting a track doesn't change output volume of other tracks
The previous change was misguided, because having the output be as
loud as possible is not necessarily a good idea, and it also meant
that the volume of each track would change when a different track
was muted, which is weird.

A side benefit of this change is that it makes the implementation
a little simpler.