Adding/removing tracks has smaller impact on volume of existing tracks
Previously, to avoid clipping the volume of each track was constrained
to be no more than (1 / trackCount). However, this meant that adding a
track would cause the existing tracks to become quieter, and removing
a track would cause the remaining tracks to become louder. This was
especially pronounced when there were a large number of tracks.

Now each track's volume will be scaled as if there were at least 8
tracks, no matter how many tracks there actually are. When there are
8 or fewer tracks, adding/removing a track won't cause any of the
existing tracks to change volume. When there are more, tracks will
change volume like before, but the relative change will be much smaller
and in theory less noticeable.