Showing computed value for attack target, not relative value
This seems maybe more clear, because it tells you exactly which
cutoff frequency the attack will end at? Although, it could also
be less clear because the range of the slider changes depending on
the value of the main cutoff frequency slider. Will try it out and
see how well it works in practice.