Offline audio generation no longer tries to prevent clipping
Instead, clipping prevention is the responsibility of the main
online audio generation. The benefit is that there is now a single
path that audio gets generated, whether online or offline. This
was mostly true before, but should now be 100% true. Besides
making the implementation simpler, it allows for a "what you hear
is what you get" system. I.e., what you hear during online playback
is the same as what you get when you download to *.wav.