*.wav download uses more exact way of determining playback time
Notes that occur at the end of a song can extend past the nominal
end point of the song, if they have a non-zero release time in their
loudness envelope. When calculating the length of time that a new
*.wav file should be, the actual point in time at which all notes
are done (including their release portion of the envelope) is now
calculated instead of using a hard-coded value.

The benefit is that this avoids either extra silence at the end of
the file, or notes being cut off before they are complete.