Extracting `BufferCollection` out of `Transport`
`Transport` should ideally be focused on scheduling playback, but
not necessarily being directly responsible for actual audio generation.
This is an initial step toward decomposing `Transport` so that it
is more focused.

The logic to generate the white noise and pink noise buffers probably
doesn't belong in the main app initialization logic, but keeping it
there for the time being until a better location is determined.