Preventing "stuck" keys on piano keyboard due to touch events
When moving fingers on the piano keyboard quickly, more than one
note could be ended at the same time. Due to a bug in how ended notes
were handled, this could result in trying to stop the same note
more than once, which caused the key to become "stuck" and not
responsive to future touches. It sometimes also caused note audio to
be held forever, until the page was refreshed.

The cause of the error was that array indices were used to remove
elements from an array, but removing an element caused the indices
to become out of sync with the array. When only one element was
removed this was OK, but it caused a problem when more than one element
was removed.