Built app is placed in /build folder, instead of top-level
This allows having the artifacts that should be deployed in a single
place, without having to remember which files should be manually
copied to the production server, and which ones should not be.

All of this is done via Webpack, except for the CSS minification
which is still separate from the time being.