Improved measure count editing; now happens in a separate mode
This avoids changing the measure count dynamically while you type,
which can cause bad things to happen. You normally only want to
set the measure could to the final value, not an intermediate value.
For example, the "natural" way to change the count from 12 -> 16
would be to erase the 2, and then add the 6. However, this would
temporarily set the measure count to 1, resulting in all but the
first measure being erased.