Adding basic support for MIDI controllers
Unfortunately this only currently works in Chrome, because that is
the only browser that currently supports WebMIDI. This also doesn't
expose any info about the currently connected MIDI controller in the
UI, so you just have to trust that plugging in your MIDI device
will work.

However, this does support highlighting notes on the piano keyboard
to match what is played on the MIDI controller, and also supports
entering notes in the pattern editor via the controller as well.