`Transport` no longer responsible for global master volume
Instead, `AudioSource` is responsible for this. `Transport` should
really be focused on current playback position, while `AudioSource`
should be responsible for mixing concerns, which includes the global
master volume.

Ultimately this is an introductory step toward converting `AudioSource`
to be a mixer object, rather than a "get your AudioContext here" object.